cpu: Move `arm` submodule to its own file.
